+++ src/main.js
@@ -2,6 +2,10 @@
 import Vue from 'vue'
 import App from './App'
 import router from './router'
+import TopHeader from './components/TopHeader/TopHeader'
+
+//注册全局组件
+Vue.component('TopHeader', TopHeader)
 
 /* eslint-disable no-new*/
 new Vue ({
