+++ src/pages/MSite/MSite.vue
@@ -1,17 +1,14 @@
 <template>
   <section class="msite">
     <!--首页头部-->
-    <header class="header">
-      <span class="header_search">
+    <TopHeader title="昌平区北七家宏福科技园(337省道北)">
+      <span class="header_search" slot="left">
         <i class="iconfont icon-sousuo"></i>
       </span>
-      <span class="header_title">
-        <span class="header_title_text ellipsis">昌平区北七家宏福科技园(337省道北)</span>
-      </span>
-      <span class="header_login">
+      <span class="header_login" slot="right">
         <span class="header_login_text">登录 | 注册</span>
       </span>
-    </header>
+    </TopHeader>
     <!--首页导航-->
     <nav class="msite_nav border-1px">
       <div class="swiper-container">
@@ -323,46 +320,6 @@
   @import "../../common/stylus/mixins.styl"
   .msite  //首页
     width 100%
-    .header
-      background-color #02a774
-      position fixed
-      z-index 100
-      left 0
-      top 0
-      width 100%
-      height 45px
-      .header_search
-        position absolute
-        left 15px
-        top 50%
-        transform translateY(-50%)
-        width 10%
-        height 50%
-        .icon-sousuo
-          font-size 25px
-          color #fff
-      .header_title
-        position absolute
-        top 50%
-        left 50%
-        transform translate(-50%, -50%)
-        width 50%
-        color #fff
-        text-align center
-        margin-left -5%
-        .header_title_text
-          font-size 20px
-          color #fff
-          display block
-      .header_login
-        font-size 14px
-        color #fff
-        position absolute
-        right 15px
-        top 50%
-        transform translateY(-50%)
-        .header_login_text
-          color #fff
     .msite_nav
       bottom-border-1px(#e4e4e4)
       margin-top 45px
